+++ pkg/api/ord/structs.go
@@ -7,22 +7,22 @@ type Entry struct {
 }
 
 type Details struct {
-	Block        int64  `json:"block"`
-	Burned       int64  `json:"burned"`
-	Divisibility int64  `json:"divisibility"`
-	Etching      string `json:"etching"`
-	Mints        int64  `json:"mints"`
-	Number       int64  `json:"number"`
-	Premine      int64  `json:"premine"`
-	SpacedRune   string `json:"spaced_rune"`
-	Symbol       string `json:"symbol"`
-	Terms        Terms  `json:"terms"`
-	Timestamp    int64  `json:"timestamp"`
-	Turbo        bool   `json:"turbo"`
+	Block        int64   `json:"block"`
+	Burned       int64   `json:"burned"`
+	Divisibility int64   `json:"divisibility"`
+	Etching      string  `json:"etching"`
+	Mints        int64   `json:"mints"`
+	Number       int64   `json:"number"`
+	Premine      float64 `json:"premine"`
+	SpacedRune   string  `json:"spaced_rune"`
+	Symbol       string  `json:"symbol"`
+	Terms        Terms   `json:"terms"`
+	Timestamp    int64   `json:"timestamp"`
+	Turbo        bool    `json:"turbo"`
 }
 
 type Terms struct {
-	Amount int64         `json:"amount"`
+	Amount float64       `json:"amount"`
 	Cap    uint64        `json:"cap"`
 	Height []int         `json:"height"`
 	Offset []interface{} `json:"offset"` // Use interface{} for mixed types (null, int)
